+++ profile_writer.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 	"os/user"
+	"path"
 	"path/filepath"
 	"time"
 
@@ -20,42 +21,51 @@ type ProfileWriter struct {
 }
 
 func NewProfileWriter() (*ProfileWriter, error) {
-	usr, err := user.Current()
-	if err != nil {
-		return nil, fmt.Errorf("Error fetching home dir: %s", err)
+	if credentialsPath, err := getCredentialsPath(); err != nil {
+		return nil, err
+	} else {
+		awsPath, _ := path.Split(credentialsPath)
+		return &ProfileWriter{
+			lock: lockfile.New(),
+			awsPath: awsPath,
+			credentialsPath: credentialsPath,
+			lockPath: filepath.Join(os.TempDir(), "swamp.lock"),
+		}, nil
 	}
+}
 
-	awsPath := filepath.Join(usr.HomeDir, ".aws")
-	credentialsPath := filepath.Join(awsPath, "credentials")
-	lockPath := filepath.Join(awsPath, ".credentials.lock")
-
-	return &ProfileWriter{
-		lock: lockfile.New(),
-		awsPath: awsPath,
-		credentialsPath: credentialsPath,
-		lockPath: lockPath,
-	}, nil
+func getCredentialsPath() (string, error) {
+	credentialsPath := os.Getenv("AWS_SHARED_CREDENTIALS_FILE")
+	if credentialsPath == "" {
+		usr, err := user.Current()
+		if err != nil {
+			return "", fmt.Errorf("Error fetching home dir: %s", err)
+		}
+		return filepath.Join(filepath.Join(usr.HomeDir, ".aws"), "credentials"), nil
+	} else {
+		return credentialsPath, nil
+	}
 }
 
 func (pw *ProfileWriter) WriteProfile(cred *sts.Credentials, profileName, region *string) error {
-	if err := pw.acquire_lock(); err != nil {
-		return err
-	}
+	pw.acquire_lock()
+	defer pw.release_lock()
 
-	cfg := pw.getOrCreateCredentialsFile()
-	if sec, err := pw.getOrCreateSection(cfg, profileName); err != nil {
+	if cfg, err := pw.getOrCreateCredentialsFile(); err != nil {
 		return err
 	} else {
-		if err := pw.writeSection(sec, cred, region); err != nil {
+		if sec, err := pw.getOrCreateSection(cfg, profileName); err != nil {
 			return err
-		}
+		} else {
+			if err := pw.writeSection(sec, cred, region); err != nil {
+				return err
+			}
 
-		if err := cfg.SaveTo(pw.credentialsPath); err != nil {
-			return fmt.Errorf("Error writing credentials file: %s", err)
+			if err := cfg.SaveTo(pw.credentialsPath); err != nil {
+				return fmt.Errorf("Error writing credentials file: %s", err)
+			}
 		}
 	}
-
-	pw.release_lock()
 
 	fmt.Printf("Wrote session token for profile %s\n", *profileName)
 	fmt.Printf("Token is valid until: %v\n", cred.Expiration)
@@ -63,35 +73,34 @@ func (pw *ProfileWriter) WriteProfile(cred *sts.Credentials, profileName, region
 	return nil
 }
 
-func (pw *ProfileWriter) acquire_lock() error {
-	if _, err := os.Stat(pw.awsPath); err != nil {
-		if err := os.MkdirAll(pw.awsPath, os.ModePerm); err != nil {
-			return fmt.Errorf("Error creating aws config path %s: %s", pw.awsPath, err)
-		}
-	}
-
+func (pw *ProfileWriter) acquire_lock() {
 	for {
 		if err := pw.lock.Lock(pw.lockPath); err == nil {
-			return nil
+			return
 		} else {
 			fmt.Printf("Waiting for lock %s\n", pw.lockPath)
 			time.Sleep(time.Second)
 		}
 	}
-	return nil
 }
 
 func (pw *ProfileWriter) release_lock() {
 	os.Remove(pw.lockPath)
 }
 
-func (pw *ProfileWriter) getOrCreateCredentialsFile() *ini.File {
+func (pw *ProfileWriter) getOrCreateCredentialsFile() (*ini.File, error) {
+	if _, err := os.Stat(pw.awsPath); err != nil {
+		if err := os.MkdirAll(pw.awsPath, os.ModePerm); err != nil {
+			return nil, fmt.Errorf("Error creating aws config path %s: %s", pw.awsPath, err)
+		}
+	}
+
 	cfg, err := ini.Load(pw.credentialsPath)
 	if err != nil {
 		fmt.Printf("Unable to find credentials file %s. Creating new file.\n", pw.credentialsPath)
 		cfg = ini.Empty()
 	}
-	return cfg
+	return cfg, nil
 }
 
 func (pw *ProfileWriter) getOrCreateSection(cfg *ini.File, profileName *string) (*ini.Section, error) {